chore(dal,sdf): split property editor module
PropertyEditor:
- Split property editor dal module into three submodules reflecting the
  three types of property editor views sent to the frontend
- No longer assemble attribute read contexts since only the component id
  is required

Component:
- Move resource-related structs and methods to a submodule
- Move "connected sockets" query to component and remove attribute
  prototypes join from query
